enacom_mcp/pyproject.toml
[project]
name = "enacom-mcp"
version = "0.1.0"
description = "MCP server for automating ENACOM Hertz DDJJ submissions via Playwright"
authors = [{name = "Osvaldo Demo", email = "demoosvaldo@gmail.com"}]
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "mcp>=1.2.0",
    "playwright>=1.45.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[project.scripts]
enacom-mcp = "enacom_mcp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["enacom_mcp"]
